store/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
psycopg2-binary==2.8.4
pytz==2020.1
sqlparse==0.3.1
